FEATURE: [!!!] Add Searchable::getMatcherValue()
Adds a Searchable::getMatcherValue() method for retrieving a specific
matcher value form a searchable using a key. The matcher value has to be
registered for that key before. For doing so, the
AbstractSearchable::registerMatcherValue() method has been added.
The AbstractSearchable::$matcherValues property has been made private
so that one has to use registerMatcherValue() when extending that class.
Additionally to `$matcherValues` all other properties in all classes that
don't have to be accessible from subclasses for the purpose of the class's
API are made private too. Use the corresponding getters and setters instead
to access these properties in subclasses.